+++ src/usr.sbin/amd/amd/util.c
@@ -242,10 +242,10 @@ char **chp;
  * addr is in network byte order.
  * sizeof(buf) needs to be at least 16.
  */
-char *inet_dquad P((char *buf, unsigned long addr));
+char *inet_dquad P((char *buf, u_int32_t addr));
 char *inet_dquad(buf, addr)
 char *buf;
-unsigned long addr;
+u_int32_t addr;
 {
 	addr = ntohl(addr);
 	sprintf(buf, "%d.%d.%d.%d",
